(refactor) Update viewResults to be passed a part's result
